questions for providers
