+++ lib/new_entry_service.rb
@@ -1,5 +1,6 @@
 require 'uri'
 require "#{(__FILE__)}/../tfa/tfa"
+require "#{(__FILE__)}/../tfa/provider"
 require "#{(__FILE__)}/../tfa/helpers"
 
 class NewEntryService
@@ -7,45 +8,11 @@ class NewEntryService
   include ::TFA::Helpers
 
   def run
-    params =
-      if question('Is the site you want to add a provider(P) or website(W)?', 'P', 'W')
-        qtfa
-      else
-        ::TFA::TFA.new.run
-      end
-  end
-
-  private
-
-  def output_yaml(params)
-
-    file = "_data/providers.yml"
-    config=YAML.load_file(file)
-
-    provider_config = {
-      "name"     => params["name"],
-      "url"      => params["url"],
-      "sms"      => params["sms"],
-      "phone"    => params["phone"],
-      "email"    => params["email"],
-      "hardware" => params["hardware"],
-      "software" => params["software"],
-    }
-
-    #remove nil values
-    provider_config.reject!{|k,v|v.nil?}
-
-    providers = config["providers"]
-
-    #append the new provider details
-    providers << provider_config
-
-    sort!(providers)
-
-    File.open(file, 'w') do |file|
-      file.write(config.to_yaml(indentation:6))
+    if question('Is the site you want to add a provider(P) or website(W)?', 'P', 'W')
+      ::TFA::Provider.new.run
+    else
+      ::TFA::TFA.new.run
     end
-
   end
 
 end
