+++ lib/tfa/tfa.rb
@@ -103,11 +103,4 @@ class TFA::TFA
 
   end
 
-  #sort the hash inplace
-  def sort!(hash)
-    hash.sort! do |x, y|
-      x["name"].downcase <=> y["name"].downcase
-    end
-  end
-
 end
